Remove unnecessary wildcard suffix
The wildcard suffix is useful for dealing with images that can only be
rendered as PNG (e.g. in the LaTeX builder); however, README.rst isn't
actually part of the docs, so it doesn't matter if the LaTeX builder
has problems with SVGs. Moreover, Github doesn't know how to handle
the wildcard, so the image is just broken on the main repo page.